Treat removable UFS card as SD card
Removable UFS cards are alternative to SD cards and they are detected as
SCSI block devices. Currently these cards shows up "USB disk" as we
consider anything other than SD card as USB disks. As both UFS cards
and USB disks are SCSI block devices, we can't differentiate between
USB disks and UFS cards by MAJOR number of block device. This change
looks for "ufs" string in DEVPATH parameter of uevent to detect if it
is UFS card or not and treat it as SD card.

Change-Id: I88534190d0610ced515c8c3e53aaae1a49480fbf
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>